--- .evergreen/config_generator/components/funcs/lint.py
@@ -1,20 +1,0 @@
-from config_generator.etc.function import Function
-from config_generator.etc.utils import bash_exec
-
-from shrub.v3.evg_command import EvgCommandType
-
-class Lint(Function):
-    name = 'lint'
-    commands = bash_exec(
-        command_type=EvgCommandType.TEST,
-        working_dir='mongo-cxx-driver',
-        script='python etc/clang_format.py lint'
-    )
-
-    @classmethod
-    def call(cls, **kwargs):
-        return cls.default_call(**kwargs)
-
-
-def functions():
-    return Lint.defn()
